sh: support -n (dry run) which just prints messages

+++ lib/formats/sh.sh
@@ -6,43 +6,60 @@
 USER='%s'
 GROUP='%s'
 
+# Set to true on -n
+DO_DRY_RUN=
+
 ##################################
 # GENERAL LINT HANDLER FUNCTIONS #
 ##################################
 
 handle_emptyfile() {
     echo 'Deleting empty file:' "$1"
-    rm -f "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        rm -f "$1"
+    fi
 }
 
 handle_emptydir() {
     echo 'Deleting empty directory:' "$1"
-    rmdir "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        rmdir "$1"
+    fi
 }
 
 handle_bad_symlink() {
     echo 'Deleting symlink pointing nowhere:' "$1"
-    rm -f "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        rm -f "$1"
+    fi
 }
 
 handle_unstripped_binary() {
     echo 'Stripping debug symbols of:' "$1"
-    strip -s "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        strip -s "$1"
+    fi
 }
 
 handle_bad_user_id() {
     echo 'chown' "$USER" "$1"
-    chmod "$USER" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        chmod "$USER" "$1"
+    fi
 }
 
 handle_bad_group_id() {
     echo 'chgrp' "$GROUP" "$1"
-    chgrp "$GROUP" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        chgrp "$GROUP" "$1"
+    fi
 }
 
 handle_bad_user_and_group_id() {
     echo 'chmod' "$USER:$GROUP" "$1"
-    chmod "$USER:$GROUP" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        chmod "$USER:$GROUP" "$1"
+    fi
 }
 
 ###############################
@@ -51,22 +68,28 @@ handle_bad_user_and_group_id() {
 
 cp_hardlink() {
     echo 'Hardlinking to original:' "$1"
-    cp --remove-destination --archive --link "$2" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        cp --remove-destination --archive --link "$2" "$1"
+    fi
 }
 
 cp_symlink() {
     echo 'Symlinking to original:' "$1"
-    touch -mr "$1" "$0"
-    cp --remove-destination --archive --symbolic-link "$2" "$1"
-    touch -mr "$0" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        touch -mr "$1" "$0"
+        cp --remove-destination --archive --symbolic-link "$2" "$1"
+        touch -mr "$0" "$1"
+    fi
 }
 
 cp_reflink() {
     # reflink $1 to $2's data, preserving $1's  mtime
     echo 'Reflinking to original:' "$1"
-    touch -mr "$1" "$0"
-    cp --reflink=always "$2" "$1"
-    touch -mr "$0" "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        touch -mr "$1" "$0"
+        cp --reflink=always "$2" "$1"
+        touch -mr "$0" "$1"
+    fi
 }
 
 skip_reflink() {
@@ -80,7 +103,9 @@ user_command() {
 
 remove_cmd() {
     echo 'Deleting:' "$1"
-    rm -rf "$1"
+    if [ -z "$DO_DRY_RUN" ]; then
+        rm -rf "$1"
+    fi
 }
 
 ##################
@@ -89,30 +114,43 @@ remove_cmd() {
 
 ask() {
     cat << EOF
+
 This script will delete certain files rmlint found.
 It is highly advisable to view the script first!
 
-Execute this script with -d to disable this message
-Hit enter to continue; CTRL-C to abort immediately
+Rmlint was executed in the following way:
+
+   $ %s
+
+Execute this script with -d to disable this informational message.
+Type any string to continue; CTRL-C, Enter or CTRL-D to abort immediately
 EOF
-    read dummy_var
+    read eof_check
+    if [ -z "$eof_check" ]
+    then
+        # Count Ctrl-D and Enter as aborted too.
+        echo "Aborted on behalf of the user."
+        exit 1;
+    fi
 }
 
 usage() {
     cat << EOF
-usage: $0 options
+usage: $0 OPTIONS
 
 OPTIONS:
--h      Show this message.
--d      Do not ask before running.
--x      Keep rmlint.sh; do note autodelete it.
+
+  -h   Show this message.
+  -d   Do not ask before running.
+  -x   Keep rmlint.sh; do not autodelete it.
+  -n   Do not perform any modifications, just print what would be done.
 EOF
 }
 
 DO_REMOVE=
 DO_ASK=
 
-while getopts "dhx" OPTION
+while getopts "dhxn" OPTION
 do
   case $OPTION in
      h)
@@ -125,6 +163,8 @@ do
      x)
        DO_REMOVE=false
        ;;
+     n)
+       DO_DRY_RUN=true
   esac
 done
